finalize starting configuration for skynet nextcloud
This now works out of the box for me.

Usage: import the module, enable the skynet service.

important points to note:

- nextcloud derivation is built into the service.  Bad design, but allows to
customize installation without an activation script having to copy over the whole
installation
- the service with the very special name is started before nginx and bind-mounts
the actual persistent store (/var/db/nextcloud) into the store.
- service dependencies are probably not complete yet, so make sure the database
is started when the webserver is running
- there is only one user name.  This would normally be bad, but this expression
is eventually supposed to run inside a nix container, so this should not matter
- when first running, the install wizard in the web will still ask for database
settings.  If preconfigured, database creation seems to fail.  In effect, the
same values from the embedded config.php have to be re-entered in the web
interface.